WIP: rbd-ggate FreeBSD libnbd guard, ISA-L build fix, cerr->std::cerr

+++ src/test/common/test_mini_flat_map.cc
@@ -38,7 +38,7 @@ struct Key {
   Key(int8_t k) : k(k) {
   }
 
-  explicit constexpr operator int8_t() const {
+  explicit constexpr operator int() const {
     return k;
   }
   Key &operator++() {
